#24 Revert back to using Long as ID instead of BigInteger.

+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/controllers/AnswerController.java
@@ -7,7 +7,7 @@ package com.gkenna.tullamoreqa.core.api.controllers;
 import com.gkenna.tullamoreqa.domain.Answer;
 import org.springframework.http.ResponseEntity;
 
-import java.math.BigInteger;
+
 
 /**
  * API Controller for the {@link Answer} Entity. This API will allow
@@ -39,7 +39,7 @@ public interface AnswerController {
      * @param answerId The ID of the {@link Answer} to retrieve.
      * @return The Response of this Request.
      */
-    ResponseEntity<Answer> getAnswer(final BigInteger answerId);
+    ResponseEntity<Answer> getAnswer(final Long answerId);
 
     /**
      * HTTP PUT Method
@@ -51,7 +51,7 @@ public interface AnswerController {
      *                 new values for answerId to update to.
      * @return The Response of this Request.
      */
-    ResponseEntity<?> updateAnswer(final BigInteger answerId,
+    ResponseEntity<?> updateAnswer(final Long answerId,
                                    final Answer input);
 
     /**
@@ -62,5 +62,5 @@ public interface AnswerController {
      * @param answerId The ID of the {@link Answer} to delete.
      * @return The Response of this Request.
      */
-    ResponseEntity<?> deleteAnswer(final BigInteger answerId);
+    ResponseEntity<?> deleteAnswer(final Long answerId);
 }
