+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import com.gkenna.tullamoreqa.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+
 
 /**
  * Repository for containing {@link User}s. Will be called by the responsible
@@ -19,7 +19,7 @@ import java.math.BigInteger;
  * @since 0.0.0
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     /**
      * Return a {@link User} based on its Username.
      *
