+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/services/EntryService.java
@@ -6,7 +6,7 @@ package com.gkenna.tullamoreqa.core.api.services;
 
 import com.gkenna.tullamoreqa.core.api.exceptions.EntryNotFoundException;
 
-import java.math.BigInteger;
+
 
 /**
  * API for interacting with {@link com.gkenna.tullamoreqa.domain.Entry}
@@ -27,7 +27,7 @@ public interface EntryService {
      * @throws EntryNotFoundException Thrown if ID of the
      *                                Entry isn't found.
      */
-    void addUpvote(final BigInteger entryId) throws EntryNotFoundException;
+    void addUpvote(final Long entryId) throws EntryNotFoundException;
 
     /**
      * Decrease the {@link com.gkenna.tullamoreqa.domain.Entry} total Upvotes
@@ -38,7 +38,7 @@ public interface EntryService {
      * @throws EntryNotFoundException Thrown if ID of the
      *                                Entry isn't found.
      */
-    void removeUpvote(final BigInteger entryId) throws EntryNotFoundException;
+    void removeUpvote(final Long entryId) throws EntryNotFoundException;
 
     /**
      * Increase the {@link com.gkenna.tullamoreqa.domain.Entry} total Downvotes
@@ -49,7 +49,7 @@ public interface EntryService {
      * @throws EntryNotFoundException Thrown if ID of the
      *                                Entry isn't found.
      */
-    void addDownvote(final BigInteger entryId) throws EntryNotFoundException;
+    void addDownvote(final Long entryId) throws EntryNotFoundException;
 
     /**
      * Decrease the {@link com.gkenna.tullamoreqa.domain.Entry} total Downvotes
@@ -60,5 +60,5 @@ public interface EntryService {
      * @throws EntryNotFoundException Thrown if ID of the
      *                                Entry isn't found.
      */
-    void removeDownvote(final BigInteger entryId) throws EntryNotFoundException;
+    void removeDownvote(final Long entryId) throws EntryNotFoundException;
 }
